% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_C.R
\name{CVaR}
\alias{CVaR}
\title{CVaR}
\usage{
CVaR(X, alpha)
}
\arguments{
\item{X}{the random sample of loss}

\item{alpha}{confidence level}
}
\value{
CVaR of loss
}
\description{
A risk measure to calculate CVaR of loss.
}
\examples{
\dontrun{
X=rnorm(1e3,5e-4,1e-3)
alpha=.05
CVaR=CVaR(X,alpha)
}
}
