% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Metropolis_c}
\alias{Metropolis_c}
\title{A Metropolis sampler using Rcpp}
\usage{
Metropolis_c(N, x0, sigma)
}
\arguments{
\item{N}{the number of samples}

\item{x0}{initial value}

\item{sigma}{variance of samples}
}
\value{
a random sample of size \code{N}
}
\description{
A Metropolis sampler using Rcpp.
}
\examples{
\dontrun{
N=1e3
sigma=2
x0=25
rw=Metropolis_c(N,x0,sigma)
}
}
