\name{StatComp20080-package}
\alias{StatComp20080-package}
\alias{StatComp20080}
\docType{package}
\title{
  Homework and project of Statistics Computation Course
}
\description{
  R versions of VaR and CVaR function are provided in this package (\code{VaR} and \code{CVaR} ). And this package also provide Rcpp versions of Metropolis function (\code{Metropolis}).
}
\author{
Xuechen Yan

Maintainer: Xuechen Yan <yanxc@163.com>
}

\keyword{ package }
