% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_V.R
\name{VaR}
\alias{VaR}
\title{VaR}
\usage{
VaR(X, alpha)
}
\arguments{
\item{X}{the random sample of loss}

\item{alpha}{confidence level}
}
\value{
VaR of loss
}
\description{
A risk measure to calculate VaR of loss.
}
\examples{
\dontrun{
X=rnorm(1e3,5e-4,1e-3)
alpha=.05
VaR=VaR(X,alpha)
}
}
